Move all ZMQ specific code into a ZMQ driver
This change moves all of our ZMQ code into a specific driver which will
allow us to create new drivers for different messaging systems to meet
operator needs and requirements.
